+++ CHANGELOG.md
@@ -5,12 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [4.0.0] - 2022-09-23
 ### Added
-- `MakeDataTransferObjectFactoryCommand` by [@regnerisch](https://github.com/regnerisch) reported by [@dimzeta](https://github.com/dimzeta) in [#66](https://github.com/regnerisch/laravel-beyond/issues/66)
+- `MakeDataTransferObjectFactoryCommand` by [@regnerisch](https://github.com/regnerisch) reported by [@thewebartisan7](https://github.com/thewebartisan7) in [#58](https://github.com/regnerisch/laravel-beyond/pull/58)
 - Extend the default `Controller` of Laravel by [@regnerisch](https://github.com/regnerisch) reported by [@thewebartisan7](https://github.com/thewebartisan7) in [#63](https://github.com/regnerisch/laravel-beyond/issues/63)
 
 ### Changed
-- Reset folder structure by [@regnerisch](https://github.com/regnerisch) inspired by [#56](https://github.com/regnerisch/laravel-beyond/issues/56)
+- Reset folder structure by [@regnerisch](https://github.com/regnerisch) reported by [@thewebartisan7](https://github.com/thewebartisan7) in [#56](https://github.com/regnerisch/laravel-beyond/issues/56)
 - Use `regnerisch/laravel-command-hooks` instead of custom `BaseCommand` by [@regnerisch](https://github.com/regnerisch)
 - Rename `--overwrite` to `--force` to be more Laravel compatible by [@regnerisch](https://github.com/regnerisch)
 
@@ -45,7 +47,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - Remove auto generation of changelog in release it by [@regnerisch](https://github.com/regnerisch)
 
-[Unreleased]: https://github.com/regnerisch/laravel-beyond/compare/v3.2.1...HEAD
+[Unreleased]: https://github.com/regnerisch/laravel-beyond/compare/v4.0.0...HEAD
+[4.0.0]: https://github.com/regnerisch/laravel-beyond/compare/v3.2.1...v4.0.0
 [3.2.1]: https://github.com/regnerisch/laravel-beyond/compare/v3.2.0...v3.2.1
 [3.2.0]: https://github.com/regnerisch/laravel-beyond/compare/v3.1.1...v3.2.0
 [3.1.1]: https://github.com/regnerisch/laravel-beyond/compare/v3.1.0...v3.1.1